requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-celery==3.3.1
django-extensions==2.2.5
django-polymorphic==2.1.2
django-rest-auth==0.9.5
django-rest-passwordreset==1.0.0
django-reverse-admin==2.7.3
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
idna==2.8
kombu==3.0.37
Markdown==3.1.1
martor==1.4.4
Pillow==6.2.1
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
